+++ Driver/io/can/can.hpp
@@ -21,7 +21,7 @@ public:
 
     bool sendFrame(const std::vector<uint8_t>& data, unsigned int id);
     // TODO: 接受指定 canID 数据帧
-    bool receiveFrame(std::vector<uint8_t>& data, std::shared_ptr<size_t> id = nullptr);
+    bool receiveFrame(std::vector<uint8_t>& data, std::shared_ptr<uint32_t> id = nullptr);
     bool isOpen() const;
 
 private:
