+++ Driver/motor/motor.hpp
@@ -31,8 +31,8 @@ public:
     Motor(
         std::shared_ptr<qdriver::interface::Interface> interfacePtr,
         const std::string& name = "defaultmotor",
-        size_t sendCanID        = 0x400,
-        size_t receiveCanID     = 0x500,
+        uint32_t sendCanID      = 0x400,
+        uint32_t receiveCanID   = 0x500,
         std::shared_ptr<qdriver::logger::Logger> logger =
             std::make_shared<qdriver::logger::Logger>()
     );
@@ -43,9 +43,9 @@ public:
 
     std::shared_ptr<qdriver::interface::Interface> getInterface(const ioType ioType) const;
 
-    bool setCanID(size_t sendCanID, size_t receiveCanID);
+    bool setCanID(uint32_t sendCanID, uint32_t receiveCanID);
 
-    bool getCanID(size_t& sendCanID, size_t& receiveCanID) const;
+    bool getCanID(uint32_t& sendCanID, uint32_t& receiveCanID) const;
 
     // 查询指令
     /**
@@ -241,8 +241,8 @@ public:
     double maxAngle_ { 2 * interface::PI };
 
 private:
-    size_t sendCanID_;
-    size_t receiveCanID_;
+    uint32_t sendCanID_;
+    uint32_t receiveCanID_;
 
     std::array<std::shared_ptr<qdriver::interface::Interface>, 2> interfaces_;
     std::shared_ptr<qdriver::logger::Logger> logger_;
